+++ borgapi/options.py
@@ -16,6 +16,16 @@ class _DefaultField:
 
 @dataclass
 class OptionsBase:
+<<<<<<< HEAD
+    """Holds all the shared methods for the subclasses
+
+    Every subclass should use this __init__ method becuase it will only set the values that the
+    dataclass supports and ignore the ones not part of it. This way the same options dict can be
+    passed to every constructor and not have to worry about duplicating flags.
+    """
+
+=======
+>>>>>>> master
     def __init__(self, **kwargs):
         default = self.defaults()
         for option in kwargs: